move the map/area partitioning API from api/map to api/datatypes/map
