+++ sdk/src/steeleagle_sdk/api/datatypes/map/partitioner/partition.py
@@ -1,0 +1,10 @@
+from __future__ import annotations
+
+from shapely.geometry import Polygon
+
+
+class Partition:
+    """Abstract base: generate partitioned segments (planar coordinates, in meters)."""
+
+    def generate_partitioned_geopoints(self, polygon: Polygon) -> list[list[tuple[float, float]]]:
+        raise NotImplementedError
